Sorted out using statements and added folders.
