+++ ExerciseTracker.Speedierone/ExerciseTracker.Speedierone/UserInput.cs
@@ -1,5 +1,4 @@
-﻿using ExerciseTracker.Speedierone.Model;
-using System.Globalization;
+﻿using System.Globalization;
 
 namespace ExerciseTracker.Speedierone
 {
